refactoring for pull request
