--- attachments/test.txt
@@ -1,1 +1,0 @@
-hallihallöle ich bin müde und will nach hause
